+++ src/main/java/sonia/scm/maven/doctor/Context.java
@@ -1,0 +1,34 @@
+package sonia.scm.maven.doctor;
+
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.apache.maven.project.MavenProject;
+
+import javax.annotation.Nullable;
+import java.util.Optional;
+
+public class Context {
+
+  private final MavenProject project;
+  private final ObjectNode packageJson;
+
+  public Context(MavenProject project) {
+    this(project, null);
+  }
+
+  public Context(MavenProject project, @Nullable ObjectNode packageJson) {
+    this.project = project;
+    this.packageJson = packageJson;
+  }
+
+  public Context withPackageJson(ObjectNode packageJson) {
+    return new Context(project, packageJson);
+  }
+
+  public Optional<ObjectNode> getPackageJson() {
+    return Optional.ofNullable(packageJson);
+  }
+
+  public MavenProject getProject() {
+    return project;
+  }
+}