+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Migrate to new packages.scm-manager.org nexus ([#6](https://github.com/scm-manager/smp-maven-plugin/pull/6))
 
 ### Added
 - Goal to copy plugin.xml and force recompile if necessary ([#3](https://github.com/scm-manager/smp-maven-plugin/pull/3))